further refactor, take out readline
